+++ sjf4j/src/main/java/org/sjf4j/node/NodeKind.java
@@ -78,6 +78,7 @@ public enum NodeKind {
         if (FacadeNodes.isNode(node)) {
             return FacadeNodes.kindOf(node);
         }
+
         return NodeKind.UNKNOWN;
     }